// packages/common-components/src/Grid/types.ts
export type AlignItems =
  | "flex-start"
  | "flex-end"
  | "center"
  | "baseline"
  | "stretch"

export type JustifyContent =
  | "flex-start"
  | "flex-end"
  | "center"
  | "space-between"
  | "space-around"
  | "space-evenly"

export type FlexDirection = "row" | "column" | "row-reverse" | "column-reverse"
export type FlexWrap = "nowrap" | "wrap" | "wrap-reverse"

export type GridSize = 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10 | 11 | 12
export type SpacingSize = 1 | 2 | 3 | 4 | 5 | 6